decomposing networks module, small fixes in builders
